Extracted `Job` code out of `AppState`
